Better separate headless tests that require webgl and headed tests.
Move example-cases to headed-cases.  Move the two-map test from gl-cases to the headed-cases directory.  Change the glheadless test to use headless Chrome.  The headless tests can be run without xvfb.
