+++ tests/all.js
@@ -16,7 +16,7 @@ if (query.test === 'all' || query.test === 'gl' || query.test === undefined) {
   tests = require.context('./gl-cases', true, /.*\.js$/);
   tests.keys().forEach(tests);
 }
-if (query.test === 'all' || query.test === 'examples') {
-  tests = require.context('./example-cases', true, /.*\.js$/);
+if (query.test === 'all' || query.test === 'headed') {
+  tests = require.context('./headed-cases', true, /.*\.js$/);
   tests.keys().forEach(tests);
 }
