+++ tests/headed-cases/lines.js
