+++ tests/test-headed.js
@@ -1,9 +1,8 @@
 /**
- * Entry point for all tests in ./example-cases/*
+ * Entry point for all tests in ./headed-cases/*
  * This is here to prevent webpack from creating a seperate bundle for each
  * test case.  See:  https://github.com/webpack/karma-webpack/issues/23
  */
 
-var tests = require.context('./example-cases', true, /.*\.js$/);
+var tests = require.context('./headed-cases', true, /.*\.js$/);
 tests.keys().forEach(tests);
-
